+++ src/caregiver/EditPatient.js
@@ -2,10 +2,6 @@ import React, { useState } from 'react'
 
 const EditPatient = (props) => {
 
-
-
-
-
     return (
         <div>
